[Issue-4133] refactor: move util
